stat graph clean old librairis + optim
