+++ dashboard/src/app/modules/stat/config/statChartOptions.ts
@@ -1,7 +1,5 @@
 import { ChartOptions, TimeScale } from 'chart.js';
 
-import { chartNamesType } from '~/core/types/stat/chartNameType';
-
 export const commonOptions: ChartOptions = {
   scales: {
     xAxes: [
@@ -49,437 +47,5 @@ export const dayOptionsTime: TimeScale = {
   tooltipFormat: 'D MMM',
   displayFormats: {
     day: 'D MMM',
-  },
-};
-
-export const statChartOptions: chartNamesType & { [key: string]: ChartOptions } = {
-  tripsPerMonth: {
-    ...commonOptions,
-    scales: {
-      xAxes: [
-        {
-          ...commonOptions.scales.xAxes[0],
-          time: {
-            ...monthOptionsTime,
-          },
-        },
-      ],
-      yAxes: [
-        {
-          ...commonOptions.scales.yAxes[0],
-        },
-      ],
-    },
-    tooltips: {
-      callbacks: {
-        label(tooltipItem, data): string {
-          return ` ${tooltipItem.yLabel} trajets`;
-        },
-      },
-    },
-  },
-  tripsPerDay: {
-    ...commonOptions,
-    scales: {
-      xAxes: [
-        {
-          ...commonOptions.scales.xAxes[0],
-          time: {
-            ...dayOptionsTime,
-          },
-        },
-      ],
-      yAxes: [
-        {
-          ...commonOptions.scales.yAxes[0],
-        },
-      ],
-    },
-    tooltips: {
-      callbacks: {
-        label(tooltipItem, data): string {
-          return ` ${tooltipItem.yLabel} trajets`;
-        },
-      },
-    },
-  },
-  tripsPerDayCumulated: {
-    ...commonOptions,
-    scales: {
-      xAxes: [
-        {
-          ...commonOptions.scales.xAxes[0],
-          time: {
-            ...dayOptionsTime,
-          },
-        },
-      ],
-      yAxes: [
-        {
-          ...commonOptions.scales.yAxes[0],
-          ticks: {
-            ...commonOptions.scales.yAxes[0].ticks,
-            min: 1,
-          },
-        },
-      ],
-    },
-    tooltips: {
-      callbacks: {
-        label(tooltipItem, data): string {
-          return ` ${tooltipItem.yLabel} trajets`;
-        },
-      },
-    },
-  },
-  tripsSubsidizedPerDay: {
-    ...commonOptions,
-    scales: {
-      xAxes: [
-        {
-          ...commonOptions.scales.xAxes[0],
-          time: {
-            ...dayOptionsTime,
-          },
-        },
-      ],
-      yAxes: [
-        {
-          ...commonOptions.scales.yAxes[0],
-        },
-      ],
-    },
-    tooltips: {
-      callbacks: {
-        label(tooltipItem, data): string {
-          return ` ${tooltipItem.yLabel} trajets`;
-        },
-      },
-    },
-  },
-  tripsSubsidizedPerMonth: {
-    ...commonOptions,
-    scales: {
-      xAxes: [
-        {
-          ...commonOptions.scales.xAxes[0],
-          time: {
-            ...monthOptionsTime,
-          },
-        },
-      ],
-      yAxes: [
-        {
-          ...commonOptions.scales.yAxes[0],
-        },
-      ],
-    },
-    tooltips: {
-      callbacks: {
-        label(tooltipItem, data): string {
-          return ` ${tooltipItem.yLabel} trajets`;
-        },
-      },
-    },
-  },
-  tripsSubsidizedPerDayCumulated: {
-    ...commonOptions,
-    scales: {
-      xAxes: [
-        {
-          ...commonOptions.scales.xAxes[0],
-          time: {
-            ...dayOptionsTime,
-          },
-        },
-      ],
-      yAxes: [
-        {
-          ...commonOptions.scales.yAxes[0],
-        },
-      ],
-    },
-    tooltips: {
-      callbacks: {
-        label(tooltipItem, data): string {
-          return ` ${tooltipItem.yLabel} trajets`;
-        },
-      },
-    },
-  },
-  distancePerDay: {
-    ...commonOptions,
-    scales: {
-      xAxes: [
-        {
-          ...commonOptions.scales.xAxes[0],
-          time: {
-            ...dayOptionsTime,
-          },
-        },
-      ],
-      yAxes: [
-        {
-          ...commonOptions.scales.yAxes[0],
-        },
-      ],
-    },
-    tooltips: {
-      callbacks: {
-        label(tooltipItem, data): string {
-          return ` ${tooltipItem.yLabel} km`;
-        },
-      },
-    },
-  },
-  distancePerMonth: {
-    ...commonOptions,
-    scales: {
-      xAxes: [
-        {
-          ...commonOptions.scales.xAxes[0],
-          time: {
-            ...monthOptionsTime,
-          },
-        },
-      ],
-      yAxes: [
-        {
-          ...commonOptions.scales.yAxes[0],
-        },
-      ],
-    },
-    tooltips: {
-      callbacks: {
-        label(tooltipItem, data): string {
-          return ` ${tooltipItem.yLabel} km`;
-        },
-      },
-    },
-  },
-  distancePerDayCumulated: {
-    ...commonOptions,
-    scales: {
-      xAxes: [
-        {
-          ...commonOptions.scales.xAxes[0],
-          time: {
-            ...dayOptionsTime,
-          },
-        },
-      ],
-      yAxes: [
-        {
-          ...commonOptions.scales.yAxes[0],
-        },
-      ],
-    },
-    tooltips: {
-      callbacks: {
-        label(tooltipItem, data): string {
-          return ` ${tooltipItem.yLabel} km`;
-        },
-      },
-    },
-  },
-  carpoolersPerMonth: {
-    ...commonOptions,
-    scales: {
-      xAxes: [
-        {
-          ...commonOptions.scales.xAxes[0],
-          time: {
-            ...monthOptionsTime,
-          },
-        },
-      ],
-      yAxes: [
-        {
-          ...commonOptions.scales.yAxes[0],
-        },
-      ],
-    },
-    tooltips: {
-      callbacks: {
-        label(tooltipItem, data): string {
-          return ` ${tooltipItem.yLabel} covoitureurs`;
-        },
-      },
-    },
-  },
-  carpoolersPerDayCumulated: {
-    ...commonOptions,
-    scales: {
-      xAxes: [
-        {
-          ...commonOptions.scales.xAxes[0],
-          time: {
-            ...dayOptionsTime,
-          },
-        },
-      ],
-      yAxes: [
-        {
-          ...commonOptions.scales.yAxes[0],
-        },
-      ],
-    },
-    tooltips: {
-      callbacks: {
-        label(tooltipItem, data): string {
-          return ` ${tooltipItem.yLabel} covoitureurs`;
-        },
-      },
-    },
-  },
-  petrolPerMonth: {
-    ...commonOptions,
-    scales: {
-      xAxes: [
-        {
-          ...commonOptions.scales.xAxes[0],
-          time: {
-            ...monthOptionsTime,
-          },
-        },
-      ],
-      yAxes: [
-        {
-          ...commonOptions.scales.yAxes[0],
-        },
-      ],
-    },
-    tooltips: {
-      callbacks: {
-        label(tooltipItem, data): string {
-          return ` ${tooltipItem.yLabel} litres`;
-        },
-      },
-    },
-  },
-  petrolPerDayCumulated: {
-    ...commonOptions,
-    scales: {
-      xAxes: [
-        {
-          ...commonOptions.scales.xAxes[0],
-          time: {
-            ...dayOptionsTime,
-          },
-        },
-      ],
-      yAxes: [
-        {
-          ...commonOptions.scales.yAxes[0],
-        },
-      ],
-    },
-    tooltips: {
-      callbacks: {
-        label(tooltipItem, data): string {
-          return ` ${tooltipItem.yLabel} litres`;
-        },
-      },
-    },
-  },
-  co2PerMonth: {
-    ...commonOptions,
-    scales: {
-      xAxes: [
-        {
-          ...commonOptions.scales.xAxes[0],
-          time: {
-            ...monthOptionsTime,
-          },
-        },
-      ],
-      yAxes: [
-        {
-          ...commonOptions.scales.yAxes[0],
-        },
-      ],
-    },
-    tooltips: {
-      callbacks: {
-        label(tooltipItem, data): string {
-          return ` ${tooltipItem.yLabel} kg`;
-        },
-      },
-    },
-  },
-  co2PerDayCumulated: {
-    ...commonOptions,
-    scales: {
-      xAxes: [
-        {
-          ...commonOptions.scales.xAxes[0],
-          time: {
-            ...dayOptionsTime,
-          },
-        },
-      ],
-      yAxes: [
-        {
-          ...commonOptions.scales.yAxes[0],
-        },
-      ],
-    },
-    tooltips: {
-      callbacks: {
-        label(tooltipItem, data): string {
-          return ` ${tooltipItem.yLabel} kg`;
-        },
-      },
-    },
-  },
-  carpoolersPerVehiculePerDay: {
-    ...commonOptions,
-    scales: {
-      xAxes: [
-        {
-          ...commonOptions.scales.xAxes[0],
-          time: {
-            ...dayOptionsTime,
-          },
-        },
-      ],
-      yAxes: [
-        {
-          ...commonOptions.scales.yAxes[0],
-        },
-      ],
-    },
-    tooltips: {
-      callbacks: {
-        label(tooltipItem, data): string {
-          return ` ${tooltipItem.yLabel} personnes`;
-        },
-      },
-    },
-  },
-  carpoolersPerVehiculePerMonth: {
-    ...commonOptions,
-    scales: {
-      xAxes: [
-        {
-          ...commonOptions.scales.xAxes[0],
-          time: {
-            ...monthOptionsTime,
-          },
-        },
-      ],
-      yAxes: [
-        {
-          ...commonOptions.scales.yAxes[0],
-        },
-      ],
-    },
-    tooltips: {
-      callbacks: {
-        label(tooltipItem, data): string {
-          return ` ${tooltipItem.yLabel} personnes`;
-        },
-      },
-    },
   },
 };
