+++ dashboard/src/app/modules/stat/modules/stat-ui/components/stat-graph-view/stat-graph-view.component.html
@@ -2,5 +2,6 @@
   <div class="stat-graph-view--loading" *ngIf="statService.isLoading">
     <mat-spinner></mat-spinner>
   </div>
-  <app-stat-graph *ngIf="!statService.isLoading" [graphName]="graphName" [graphData]="graphData"></app-stat-graph>
+
+  <app-stat-graph [graphName]="graphName"></app-stat-graph>
 </div>
