+++ dashboard/src/app/modules/stat/modules/stat-ui/components/stat-graph/stat-graph-petrol/stat-graph-petrol.component.ts
@@ -3,7 +3,7 @@ import { FormatedStatInterface } from '~/core/interfaces/stat/formatedStatInterf
 import { StatInterface } from '~/core/interfaces/stat/StatInterface';
 import { petrolFactor } from '~/modules/stat/config/stat';
 import { ApiGraphTimeMode } from '~/modules/stat/services/ApiGraphTimeMode';
-import { formatMonthLabel, formatDayLabel } from '~/modules/stat/services/stat-format.service';
+import { formatDayLabel, formatMonthLabel } from '~/modules/stat/stat-format';
 import { commonOptions, monthOptionsTime, dayOptionsTime } from '../../../../../config/statChartOptions';
 
 import { GraphTimeMode, GraphTimeModeLabel } from '../../../GraphTimeMode';
