+++ dashboard/src/app/modules/stat/modules/stat-ui/components/stat-graph/stat-graph-timemode-nav/stat-graph-timemode-nav.component.html
@@ -1,4 +1,3 @@
-{{ navList }}
 <mat-button-toggle-group class="graph-time-nav" [(ngModel)]="nav" (ngModelChange)="change()">
   <mat-button-toggle *ngFor="let navElement of navList" [aria-label]="label(navElement)" [value]="navElement">
     {{ label(navElement) }}
