+++ dashboard/src/app/modules/stat/modules/stat-ui/components/stat-territory-view/stat-territory-view.component.html
@@ -1,9 +1,10 @@
 <div class="stats--loading" *ngIf="statService.isLoading">
   <mat-spinner></mat-spinner>
 </div>
-<div class="stats" *ngIf="loaded">
+
+<div class="stats">
   <div class="stats-graph">
-    <app-stat-graph [graphName]="graphName" [graphData]="graphData"></app-stat-graph>
+    <app-stat-graph [graphName]="graphName"></app-stat-graph>
   </div>
   <div *ngIf="statService.stat" class="stats-numbers">
     <app-stat-number
