+++ dashboard/src/app/modules/stat/stat-format.ts
@@ -1,0 +1,13 @@
+const monthsLabel = ['Jan', 'Fev', 'Mar', 'Avr', 'Mai', 'Jun', 'Jui', 'Sep', 'Oct', 'Nov', 'Dec'];
+
+// format yyyy-mm formated string to month label
+export function formatMonthLabel(monthDate) {
+  const monthDateSpl = monthDate.split('-');
+  return `${monthsLabel[parseInt(monthDateSpl[1]) - 1]} ${monthDateSpl[0]}`;
+}
+
+// format iso string date to graph label
+export function formatDayLabel(dayDate) {
+  const date = new Date(dayDate);
+  return `${date.getDate()} ${monthsLabel[date.getMonth()]}`;
+}
